[desafio-10] Move turing logic to own pkg
