+++ desafio-10/MatMercer/go/go.mod
@@ -1,0 +1,3 @@
+module matbm.net/turing-machine
+
+go 1.20
